+++ dbt2looker/generator.py
@@ -1,4 +1,5 @@
 import logging
+import re
 
 import lkml
 
@@ -154,6 +155,21 @@ LOOKER_DTYPE_MAP = {
         'TIME WITHOUT TIME ZONE': 'string',
         # TIMETZ not supported
         # TIME WITH TIME ZONE not supported
+    },
+    'spark': {
+        'byte':        'number',
+        'short':       'number',
+        'integer':     'number',
+        'long':        'number',
+        'float':       'number',
+        'double':      'number',
+        'decimal':     'number',
+        'string':      'string',
+        'varchar':     'string',
+        'char':        'string',
+        'boolean':     'yesno',
+        'timestamp':   'timestamp',
+        'date':        'datetime',
     }
 }
 
@@ -172,8 +188,13 @@ looker_timeframes = [
 ]
 
 
+def normalise_spark_types(column_type: str) -> str:
+    return re.match(r'^[^\(]*', column_type)
+
+
 def map_adapter_type_to_looker(adapter_type: models.SupportedDbtAdapters, column_type: str):
-    looker_type = LOOKER_DTYPE_MAP[adapter_type].get(column_type)
+    normalised_column_type = normalise_spark_types(column_type) if adapter_type == models.SupportedDbtAdapters.spark.value else column_type
+    looker_type = LOOKER_DTYPE_MAP[adapter_type].get(normalised_column_type)
     if (column_type is not None) and (looker_type is None):
         logging.warning(f'Column type {column_type} not supported for conversion from {adapter_type} to looker. No dimension will be created.')
     return looker_type
@@ -282,7 +303,7 @@ def lookml_view_from_dbt_model(model: models.DbtModel, adapter_type: models.Supp
     lookml = {
         'view': {
             'name': model.name,
-            'sql_table_name': f'{model.database}.{model.db_schema}.{model.name}',
+            'sql_table_name': model.relation_name,
             'dimension_groups': lookml_dimension_groups_from_model(model, adapter_type),
             'dimensions': lookml_dimensions_from_model(model, adapter_type),
             'measures': lookml_measures_from_model(model),